requirements.txt
django>=1.9,<1.10
djangorestframework>=3.3,<3.4
Pillow
